Multivariate bias correction (Spearman rank correlation)

Multivariate bias correction that matches marginal distributions using QDM and the Spearman rank correlation dependence structure following Cannon (2016).

MBCr(o.c, m.c, m.p, iter=20, cor.thresh=1e-4,
     ratio.seq=rep(FALSE, ncol(o.c)), trace=0.05,
     trace.calc=0.5*trace, jitter.factor=0, n.tau=NULL,
     ratio.max=2, ratio.max.trace=10*trace, ties='first',
     qmap.precalc=FALSE, silent=FALSE, subsample=NULL,
     pp.type=7)

Arguments

o.c

matrix of observed samples during the calibration period.

m.c

matrix of model outputs during the calibration period.

m.p

matrix of model outputs during the projected period.

iter

maximum number of algorithm iterations.

cor.thresh

if greater than zero, a threshold indicating the change in magnitude of Spearman rank correlations required for convergence.

ratio.seq

vector of logical values indicating if samples are of a ratio quantity (e.g., precipitation).

trace

numeric values indicating thresholds below which values of a ratio quantity (e.g., ratio=TRUE) should be considered exact zeros.

trace.calc

numeric values of thresholds used internally when handling of exact zeros; defaults to one half of trace.

jitter.factor

optional strength of jittering to be applied when quantities are quantized.

n.tau

number of quantiles used in the quantile mapping; NULL equals the length of the m.p series.

ratio.max

numeric values indicating the maximum proportional changes allowed for ratio quantities below the ratio.max.trace threshold.

ratio.max.trace

numeric values of trace thresholds used to constrain the proportional change in ratio quantities to ratio.max; defaults to ten times trace.

ties

method used to handle ties when calculating ordinal ranks.

qmap.precalc

logical value indicating if m.c and m.p are outputs from QDM.

silent

logical value indicating if algorithm progress should be reported.

subsample

use subsample draws of size n.tau to calculate empirical quantiles; if NULL, calculate normally.

pp.type

type of plotting position used in quantile.

Value

a list of with elements consisting of:

mhat.c

matrix of bias corrected m.c values for the calibration period.

mhat.p

matrix of bias corrected m.p values for the projection period.

See also

References

Cannon, A.J., 2016. Multivariate bias correction of climate model output: Matching marginal distributions and inter-variable dependence structure. Journal of Climate, 29:7045-7064. doi:10.1175/JCLI-D-15-0679.1

Cannon, A.J., S.R. Sobie, and T.Q. Murdock, 2015. Bias correction of simulated precipitation by quantile mapping: How well do methods preserve relative changes in quantiles and extremes? Journal of Climate, 28:6938-6959. doi:10.1175/JCLI-D-14-00754.1