Multivariate linear rescaling using Cholesky decomposition

Multivariate linear bias correction based on Cholesky decomposition of the covariance matrix following Scheuer and Stoller (1962) and Bürger et al. (2011). Bias correction matches the multivariate mean and covariance structure.

MRS(
  o.c,
  m.c,
  m.p,
  o.c.chol = NULL,
  o.p.chol = NULL,
  m.c.chol = NULL,
  m.p.chol = NULL
)

Arguments

o.c

matrix of observed samples during the calibration period.

m.c

matrix of model outputs during the calibration period.

m.p

matrix of model outputs during the projected period.

o.c.chol

precalculated Cholesky decomposition of the o.c covariance matrix; NULL calculates internally.

o.p.chol

precalculated Cholesky decomposition of the target o.p covariance matrix; NULL defaults to o.c.chol.

m.c.chol

precalculated Cholesky decomposition of the m.c covariance matrix; NULL calculates internally.

m.p.chol

precalculated Cholesky decomposition of the m.p covariance matrix; NULL calculates internally.

Value

a list of with elements consisting of:

mhat.c

matrix of bias corrected m.c values for the calibration period.

mhat.p

matrix of bias corrected m.p values for the projection period.

References

Scheuer, E.M. and D.S. Stoller, 1962. On the generation of normal random vectors. Technometrics, 4(2):278-281.

Bürger, G., J. Schulla, and A.T. Werner, 2011. Estimates of future flow, including extremes, of the Columbia River headwaters. Water Resources Research, 47(10):W10520. doi:10.1029/2010WR009716

See also