Multivariate bias correction (R2D2)

Multivariate bias correction that matches the multivariate distribution using QDM and the R2D2 algorithm following Vrac (2018).

R2D2(o.c, m.c, m.p, ref.column=1, ratio.seq=rep(FALSE, ncol(o.c)),
     trace=0.05, trace.calc=0.5*trace, jitter.factor=0,
     n.tau=NULL, ratio.max=2, ratio.max.trace=10*trace,
     ties='first', qmap.precalc=FALSE, subsample=NULL, pp.type=7)

Arguments

o.c

matrix of observed samples during the calibration period.

m.c

matrix of model outputs during the calibration period.

m.p

matrix of model outputs during the projected period.

ref.column

index of the reference column used for the 1D nearest neighbour matching

ratio.seq

vector of logical values indicating if samples are of a ratio quantity (e.g., precipitation).

trace

numeric values indicating thresholds below which values of a ratio quantity (e.g., ratio=TRUE) should be considered exact zeros.

trace.calc

numeric values of thresholds used internally when handling of exact zeros; defaults to one half of trace.

jitter.factor

optional strength of jittering to be applied when quantities are quantized.

n.tau

number of quantiles used in the quantile mapping; NULL equals the length of the m.p series.

ratio.max

numeric values indicating the maximum proportional changes allowed for ratio quantities below the ratio.max.trace threshold.

ratio.max.trace

numeric values of trace thresholds used to constrain the proportional change in ratio quantities to ratio.max; defaults to ten times trace.

ties

method used to handle ties when calculating ordinal ranks.

qmap.precalc

logical value indicating if m.c and m.p are outputs from QDM.

subsample

use subsample draws of size n.tau to calculate initial empirical quantiles; if NULL, calculate normally.

pp.type

type of plotting position used in quantile.

Value

a list of with elements consisting of:

mhat.c

matrix of bias corrected m.c values for the calibration period.

mhat.p

matrix of bias corrected m.p values for the projection period.

See also

References

Cannon, A.J., S.R. Sobie, and T.Q. Murdock, 2015. Bias correction of simulated precipitation by quantile mapping: How well do methods preserve relative changes in quantiles and extremes? Journal of Climate, 28:6938-6959. doi:10.1175/JCLI-D-14-00754.1

Vrac, M., 2018. Multivariate bias adjustment of high-dimensional climate simulations: the Rank Resampling for Distributions and Dependences (R2D2) bias correction. Hydrology and Earth System Sciences, 22:3175-3196. doi:10.5194/hess-22-3175-2018