Energy distance score

Calculate the energy distance score measuring the statistical discrepancy between samples x and y from two multivariate distributions.

escore(x, y, scale.x = FALSE, n.cases = NULL, alpha = 1, method = "cluster")

Arguments

x

numeric matrix.

y

numeric matrix.

scale.x

logical indicating whether data should be standardized based on x.

n.cases

the number of sub-sampled cases; NULL uses all data.

alpha

distance exponent in (0,2]

method

method used to weight the statistics

References

Székely, G.J. and M.L. Rizzo, 2004. Testing for equal distributions in high dimension, InterStat, November (5).

Székely, G.J. and M.L. Rizzo, 2013. Energy statistics: statistics based on distances. Journal of Statistical Planning and Inference, 143(8):1249-1272. doi:10.1016/j.jspi.2013.03.018

Rizzo, M.L. and G.L. Székely, 2016. Energy distance. Wiley Interdisciplinary Reviews: Computational Statistics, 8(1):27-38.